checkin/requirements.txt
aniso8601==8.0.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
flask-mongoengine==0.9.5
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
pymongo==3.11.0
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
